+++ DekorTrade/src/hu/dekortrade/server/sync/CikkSzinkron.java
@@ -10,17 +10,17 @@ public class CikkSzinkron {
 
 	private String megnevezes;
 
-	private Double ar;
+	private Float ar;
 
 	private Integer kiskarton;
 
 	private Integer darab;
 
-	private Double terfogat;
+	private Float terfogat;
 
-	private Double bsuly;
+	private Float bsuly;
 
-	private Double nsuly;
+	private Float nsuly;
 
 	private Integer kepek;
 
@@ -31,8 +31,8 @@ public class CikkSzinkron {
 	}
 
 	public CikkSzinkron(String fotipus, String altipus, String cikkszam,
-			String megnevezes, Double ar, Integer kiskarton, Integer darab,
-			Double terfogat, Double bsuly, Double nsuly, Integer kepek,
+			String megnevezes, Float ar, Integer kiskarton, Integer darab,
+			Float terfogat, Float bsuly, Float nsuly, Integer kepek,
 			Boolean torolt) {
 		this.fotipus = fotipus;
 		this.altipus = altipus;
@@ -80,11 +80,11 @@ public class CikkSzinkron {
 		this.megnevezes = megnevezes;
 	}
 
-	public Double getAr() {
+	public Float getAr() {
 		return ar;
 	}
 
-	public void setAr(Double ar) {
+	public void setAr(Float ar) {
 		this.ar = ar;
 	}
 
@@ -104,27 +104,27 @@ public class CikkSzinkron {
 		this.darab = darab;
 	}
 
-	public Double getTerfogat() {
+	public Float getTerfogat() {
 		return terfogat;
 	}
 
-	public void setTerfogat(Double terfogat) {
+	public void setTerfogat(Float terfogat) {
 		this.terfogat = terfogat;
 	}
 
-	public Double getBsuly() {
+	public Float getBsuly() {
 		return bsuly;
 	}
 
-	public void setBsuly(Double bsuly) {
+	public void setBsuly(Float bsuly) {
 		this.bsuly = bsuly;
 	}
 
-	public Double getNsuly() {
+	public Float getNsuly() {
 		return nsuly;
 	}
 
-	public void setNsuly(Double nsuly) {
+	public void setNsuly(Float nsuly) {
 		this.nsuly = nsuly;
 	}
