+++ DekorTrade99/src/hu/dekortrade99/server/sync/CikkSzinkron.java
@@ -10,17 +10,17 @@ public class CikkSzinkron {
 
 	private String megnevezes;
 
-	private Double ar;
+	private Float ar;
 
 	private Integer kiskarton;
 
 	private Integer darab;
 
-	private Double terfogat;
+	private Float terfogat;
 
-	private Double bsuly;
+	private Float bsuly;
 
-	private Double nsuly;
+	private Float nsuly;
 
 	private Integer kepek;
 
@@ -30,9 +30,9 @@ public class CikkSzinkron {
 		
 	}
 	
-	public CikkSzinkron(String fotipus, String altipus, String cikkszam, String megnevezes, Double ar,
-			Integer kiskarton, Integer darab, Double terfogat,
-			Double bsuly, Double nsuly, Integer kepek, Boolean torolt) {
+	public CikkSzinkron(String fotipus, String altipus, String cikkszam, String megnevezes, Float ar,
+			Integer kiskarton, Integer darab, Float terfogat,
+			Float bsuly, Float nsuly, Integer kepek, Boolean torolt) {
 		this.fotipus = fotipus;
 		this.altipus = altipus;
 		this.megnevezes = megnevezes;
@@ -78,11 +78,11 @@ public class CikkSzinkron {
 		this.megnevezes = megnevezes;
 	}
 
-	public Double getAr() {
+	public Float getAr() {
 		return ar;
 	}
 
-	public void setAr(Double ar) {
+	public void setAr(Float ar) {
 		this.ar = ar;
 	}
 
@@ -102,27 +102,27 @@ public class CikkSzinkron {
 		this.darab = darab;
 	}
 
-	public Double getTerfogat() {
+	public Float getTerfogat() {
 		return terfogat;
 	}
 
-	public void setTerfogat(Double terfogat) {
+	public void setTerfogat(Float terfogat) {
 		this.terfogat = terfogat;
 	}
 
-	public Double getBsuly() {
+	public Float getBsuly() {
 		return bsuly;
 	}
 
-	public void setBsuly(Double bsuly) {
+	public void setBsuly(Float bsuly) {
 		this.bsuly = bsuly;
 	}
 
-	public Double getNsuly() {
+	public Float getNsuly() {
 		return nsuly;
 	}
 
-	public void setNsuly(Double nsuly) {
+	public void setNsuly(Float nsuly) {
 		this.nsuly = nsuly;
 	}
